working Makefile
