+++ main.c
@@ -1,0 +1,117 @@
+/* 
+ * simple lennard-jones potential MD code with velocity verlet.
+ * units: Length=Angstrom, Mass=amu; Energy=kcal
+ *
+ */
+
+#include <stdio.h>
+#include <string.h>
+#include <ctype.h>
+#include <stdlib.h>
+#include <math.h>
+#include <ljmd.h>
+
+int main(int argc, char **argv) 
+{
+    int nprint, i;
+    char restfile[BLEN], trajfile[BLEN], ergfile[BLEN], line[BLEN];
+    FILE *fp,*traj,*erg;
+    mdsys_t sys;
+
+    /* read input file */
+    if(get_a_line(stdin,line)) return 1;
+    sys.natoms=atoi(line);
+    if(get_a_line(stdin,line)) return 1;
+    sys.mass=atof(line);
+    if(get_a_line(stdin,line)) return 1;
+    sys.epsilon=atof(line);
+    if(get_a_line(stdin,line)) return 1;
+    sys.sigma=atof(line);
+    if(get_a_line(stdin,line)) return 1;
+    sys.rcut=atof(line);
+    if(get_a_line(stdin,line)) return 1;
+    sys.box=atof(line);
+    if(get_a_line(stdin,restfile)) return 1;
+    if(get_a_line(stdin,trajfile)) return 1;
+    if(get_a_line(stdin,ergfile)) return 1;
+    if(get_a_line(stdin,line)) return 1;
+    sys.nsteps=atoi(line);
+    if(get_a_line(stdin,line)) return 1;
+    sys.dt=atof(line);
+    if(get_a_line(stdin,line)) return 1;
+    nprint=atoi(line);
+
+    /* allocate memory */
+    sys.rx=(double *)malloc(sys.natoms*sizeof(double));
+    sys.ry=(double *)malloc(sys.natoms*sizeof(double));
+    sys.rz=(double *)malloc(sys.natoms*sizeof(double));
+    sys.vx=(double *)malloc(sys.natoms*sizeof(double));
+    sys.vy=(double *)malloc(sys.natoms*sizeof(double));
+    sys.vz=(double *)malloc(sys.natoms*sizeof(double));
+    sys.fx=(double *)malloc(sys.natoms*sizeof(double));
+    sys.fy=(double *)malloc(sys.natoms*sizeof(double));
+    sys.fz=(double *)malloc(sys.natoms*sizeof(double));
+
+    /* read restart */
+    fp=fopen(restfile,"r");
+    if(fp) {
+        for (i=0; i<sys.natoms; ++i) {
+            fscanf(fp,"%lf%lf%lf",sys.rx+i, sys.ry+i, sys.rz+i);
+        }
+        for (i=0; i<sys.natoms; ++i) {
+            fscanf(fp,"%lf%lf%lf",sys.vx+i, sys.vy+i, sys.vz+i);
+        }
+        fclose(fp);
+        azzero(sys.fx, sys.natoms);
+        azzero(sys.fy, sys.natoms);
+        azzero(sys.fz, sys.natoms);
+    } else {
+        perror("cannot read restart file");
+        return 3;
+    }
+
+    /* initialize forces and energies.*/
+    sys.nfi=0;
+    force(&sys);
+    ekin(&sys);
+    
+    erg=fopen(ergfile,"w");
+    traj=fopen(trajfile,"w");
+
+    printf("Starting simulation with %d atoms for %d steps.\n",sys.natoms, sys.nsteps);
+    printf("     NFI            TEMP            EKIN                 EPOT              ETOT\n");
+    output(&sys, erg, traj);
+
+    /**************************************************/
+    /* main MD loop */
+    for(sys.nfi=1; sys.nfi <= sys.nsteps; ++sys.nfi) {
+
+        /* write output, if requested */
+        if ((sys.nfi % nprint) == 0)
+            output(&sys, erg, traj);
+
+        /* propagate system and recompute energies */
+        velverlet_1(&sys);
+	force(&sys);
+	velverlet_2(&sys);
+        ekin(&sys);
+    }
+    /**************************************************/
+
+    /* clean up: close files, free memory */
+    printf("Simulation Done.\n");
+    fclose(erg);
+    fclose(traj);
+
+    free(sys.rx);
+    free(sys.ry);
+    free(sys.rz);
+    free(sys.vx);
+    free(sys.vy);
+    free(sys.vz);
+    free(sys.fx);
+    free(sys.fy);
+    free(sys.fz);
+
+    return 0;
+}
